+++ cmd/api/main.go
@@ -12,11 +12,11 @@ import (
 
 	"github.com/azusachino/felicia"
 	"github.com/azusachino/felicia/apps/core/domain"
+	"github.com/azusachino/felicia/apps/runtime/importer"
 	"github.com/azusachino/felicia/internal/api"
 	"github.com/azusachino/felicia/internal/config"
 	"github.com/azusachino/felicia/internal/dawarich"
 	"github.com/azusachino/felicia/internal/immich"
-	"github.com/azusachino/felicia/internal/importer"
 	"github.com/azusachino/felicia/internal/store/pg"
 )
