+++ go.mod
@@ -4,6 +4,7 @@ go 1.26
 
 require (
 	github.com/azusachino/felicia/apps/core v0.1.0
+	github.com/azusachino/felicia/apps/runtime v0.1.0
 	github.com/google/go-cmp v0.7.0
 	github.com/google/uuid v1.6.0
 	github.com/jackc/pgx/v5 v5.10.0
