+++ internal/api/server.go
@@ -17,7 +17,7 @@ import (
 	"github.com/paulmach/orb"
 
 	"github.com/azusachino/felicia/apps/core/domain"
-	"github.com/azusachino/felicia/internal/importer"
+	"github.com/azusachino/felicia/apps/runtime/importer"
 	"github.com/azusachino/felicia/internal/publication"
 )
