+++ internal/store/pg/write_side_integration_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/paulmach/orb"
 
 	"github.com/azusachino/felicia/apps/core/domain"
-	"github.com/azusachino/felicia/internal/importer"
+	"github.com/azusachino/felicia/apps/runtime/importer"
 	"github.com/azusachino/felicia/internal/store/pg"
 )
